Fix wonky job name in BlueMessageEnricher
